+++ R/unname_all_chunks.R
@@ -1,15 +1,17 @@
 #' @title Unname chunks in a single file
 #'
-#' @description Unname all chunks except the setup chunk, in a single file
+#' @description Unname in a single file all chunks except the setup chunk, or alternatively unname the chunknames with a given prefix
 #'
 #' @inherit name_chunks details
 #'
 #' @param path Path to file
+#' @param ch_n_p Character string with prefix of chunknames that will be removed. Default: NULL (indicating all chunknames will be removed except the one named `setup`)
 #'
 #' @export
 #'
 #' @examples
-#' temp_file_path <- file.path(tempdir(), "test.Rmd")
+#' # remove all chunklabels except the one named 'setup'
+#' temp_file_path <- file.path(tempdir(), "test1.Rmd")
 #' file.copy(system.file("examples", "example4.Rmd", package = "namer"),
 #'           temp_file_path,
 #'           overwrite = TRUE)
@@ -17,7 +19,16 @@
 #' if(interactive()){
 #' file.edit(temp_file_path)
 #' }
-unname_all_chunks <- function(path){
+#' # remove all chunk labels starting with 'example4'
+#' temp_file_path <- file.path(tempdir(), "test2.Rmd")
+#' file.copy(system.file("examples", "example4.Rmd", package = "namer"),
+#'           temp_file_path,
+#'           overwrite = TRUE)
+#' unname_all_chunks(temp_file_path,ch_n_p='example4')
+#' if(interactive()){
+#' file.edit(temp_file_path)
+#' }
+unname_all_chunks <- function(path,ch_n_p=NULL){
   # read the whole file
   lines <- readLines(path)
 
@@ -29,8 +40,16 @@ unname_all_chunks <- function(path){
     return(invisible("TRUE"))
   }
 
-  # preserve the setup label, delete the others
-  chunk_headers_info$name[chunk_headers_info$name != "setup"] <- ""
+  if ( is.null(ch_n_p)){
+    # preserve the setup label, delete the others
+    chunk_headers_info$name[chunk_headers_info$name != "setup"] <- ""
+  } else {
+    # preserve labels not starting with ch_n_p
+    del_labels = stringr::str_detect(
+      stringr::str_replace_na(chunk_headers_info$name, replacement = ""),
+      glue::glue('^{ch_n_p}'))
+    chunk_headers_info$name[del_labels] <- ""
+  }
 
   newlines <- re_write_headers(chunk_headers_info)