+++ R/unname_all_chunks.R
@@ -1,11 +1,11 @@
 #' @title Unname chunks in a single file
 #'
-#' @description Unname in a single file all chunks except the setup chunk, or alternatively unname the chunknames with a given prefix
+#' @description Unname in a single file all chunks except the setup chunk, or alternatively only unname the chunknames with a given prefix
 #'
 #' @inherit name_chunks details
 #'
 #' @param path Path to file
-#' @param ch_n_p Character string with prefix of chunknames that will be removed. Default: NULL (indicating all chunknames will be removed except the one named `setup`)
+#' @param chunk_name_prefix Character string with prefix of chunknames that will be removed. Default: NULL (indicating all chunknames will be removed except the one named `setup`)
 #'
 #' @export
 #'
@@ -24,11 +24,11 @@
 #' file.copy(system.file("examples", "example4.Rmd", package = "namer"),
 #'           temp_file_path,
 #'           overwrite = TRUE)
-#' unname_all_chunks(temp_file_path,ch_n_p='example4')
+#' unname_all_chunks(temp_file_path,chunk_name_prefix='example4')
 #' if(interactive()){
 #' file.edit(temp_file_path)
 #' }
-unname_all_chunks <- function(path,ch_n_p=NULL){
+unname_all_chunks <- function(path,chunk_name_prefix=NULL){
   # read the whole file
   lines <- readLines(path)
 
@@ -40,14 +40,13 @@ unname_all_chunks <- function(path,ch_n_p=NULL){
     return(invisible("TRUE"))
   }
 
-  if ( is.null(ch_n_p)){
+  if ( is.null(chunk_name_prefix)){
     # preserve the setup label, delete the others
     chunk_headers_info$name[chunk_headers_info$name != "setup"] <- ""
   } else {
-    # preserve labels not starting with ch_n_p
-    del_labels = stringr::str_detect(
-      stringr::str_replace_na(chunk_headers_info$name, replacement = ""),
-      glue::glue('^{ch_n_p}'))
+    # preserve labels not starting with chunk_name_prefix
+    del_labels = strtrim(chunk_headers_info$name,nchar(chunk_name_prefix)) %in%
+      chunk_name_prefix
     chunk_headers_info$name[del_labels] <- ""
   }