+++ R/unname_all_chunks.R
@@ -1,15 +1,19 @@
 #' @title Unname chunks in a single file
 #'
-#' @description Unname all chunks except the setup chunk, in a single file
+#' @description Unname in a single file all chunks,
+#' or alternatively only unname the chunknames with a given prefix.
+#' In both cases, the chunk name "setup" is preserved, that chunk is never unnamed.
 #'
 #' @inherit name_chunks details
 #'
 #' @param path Path to file
+#' @param chunk_name_prefix Character string with prefix of chunknames that will be removed. Default: NULL (indicating all chunknames will be removed except the one named `setup`)
 #'
 #' @export
 #'
 #' @examples
-#' temp_file_path <- file.path(tempdir(), "test.Rmd")
+#' # remove all chunklabels except the one named 'setup'
+#' temp_file_path <- file.path(tempdir(), "test1.Rmd")
 #' file.copy(system.file("examples", "example4.Rmd", package = "namer"),
 #'           temp_file_path,
 #'           overwrite = TRUE)
@@ -17,7 +21,16 @@
 #' if(interactive()){
 #' file.edit(temp_file_path)
 #' }
-unname_all_chunks <- function(path){
+#' # remove all chunk labels starting with 'example4'
+#' temp_file_path <- file.path(tempdir(), "test2.Rmd")
+#' file.copy(system.file("examples", "example4.Rmd", package = "namer"),
+#'           temp_file_path,
+#'           overwrite = TRUE)
+#' unname_all_chunks(temp_file_path,chunk_name_prefix='example4')
+#' if(interactive()){
+#' file.edit(temp_file_path)
+#' }
+unname_all_chunks <- function(path,chunk_name_prefix=NULL){
   # read the whole file
   lines <- readLines(path)
 
@@ -29,8 +42,17 @@ unname_all_chunks <- function(path){
     return(invisible("TRUE"))
   }
 
-  # preserve the setup label, delete the others
-  chunk_headers_info$name[chunk_headers_info$name != "setup"] <- ""
+  if ( is.null(chunk_name_prefix)){
+    # preserve the setup label, delete the others
+    chunk_headers_info$name[chunk_headers_info$name != "setup"] <- ""
+  } else {
+    # preserve labels not starting with chunk_name_prefix
+    del_labels <- strtrim(chunk_headers_info$name,nchar(chunk_name_prefix)) %in%
+      chunk_name_prefix
+    setup_label <- !(chunk_headers_info$name %in% 'setup')
+    del_labels <- del_labels & setup_label
+    chunk_headers_info$name[del_labels] <- ""
+  }
 
   newlines <- re_write_headers(chunk_headers_info)