+++ NEWS.md
@@ -1,4 +1,5 @@
 # namer (development version)
+* unname_all_chunks now accepts argument `ch_n_p` with the prefix of the chunknames to be unnamed
 
 # namer 0.1.4