+++ NEWS.md
@@ -3,6 +3,7 @@
 * `name_chunks()` and `name_dir_chunks()` are now able to unname all chunks before naming them. This ensures a consistent naming for all chunks instead of just labelling unnamed chunks (@pat-s, #<>).
 * new function `unname_dir_chunks()` that works in the same way as `name_dir_chunks()` (@pat-s, #<>)
 * rename `unname_all_chunks()` to `unname_chunks()` to be consistend with `name_chunks()`
+* added some `cli` verbosity (@pat-s, #<>)
 
 # namer 0.1.4